`GlitterBridgeSDK` clean up and types updtes for algorand
