+++ .lintstagedrc.json
@@ -1,4 +1,3 @@
 {
-    "*.{js,ts,md,json,yml,yaml}": "prettier --write",
-    "*.{js,ts}": "eslint --cache --fix --no-error-on-unmatched-pattern"
+  "*.{js,ts}": "eslint --cache --fix --no-error-on-unmatched-pattern"
 }
