+++ packages/core/src/GlitterBridgeSDK.ts
@@ -1,15 +1,14 @@
-import { AlgorandConnect } from "./lib/chains/algorand";
-import { EvmConnect } from "./lib/chains/evm";
-import { SolanaConnect } from "./lib/chains/solana";
-import { TronConnect } from "./lib/chains/tron/connect";
-import { BridgeToken, BridgeTokens } from "./lib/common";
+import {AlgorandConnect} from "./lib/chains/algorand";
+import {EvmConnect} from "./lib/chains/evm";
+import {SolanaConnect} from "./lib/chains/solana";
+import {TronConnect} from "./lib/chains/tron/connect";
 import {
     BridgeEvmNetworks,
     BridgeNetworks,
 } from "./lib/common/networks/networks";
-import { GlitterBridgeConfig, GlitterEnvironment } from "./lib/configs/config";
-import { BridgeMainnet } from "./lib/configs/networks/mainnet";
-import { BridgeTestnet } from "./lib/configs/networks/testnet";
+import {GlitterBridgeConfig, GlitterEnvironment} from "./lib/configs/config";
+import {BridgeMainnet} from "./lib/configs/networks/mainnet";
+import {BridgeTestnet} from "./lib/configs/networks/testnet";
 
 /**
  * Glitter Bridge SDK
@@ -20,38 +19,25 @@ export class GlitterBridgeSDK {
     private _bridgeConfig: GlitterBridgeConfig | undefined;
     /** RPC URL Override */
     private _rpcOverrides: { [key: string]: string } = {};
-
     /** Chain Specific SDKs */
     private _evm: Map<BridgeEvmNetworks, EvmConnect | undefined> = new Map();
     private _algorand: AlgorandConnect | undefined;
     private _solana: SolanaConnect | undefined;
     private _tron: TronConnect | undefined;
 
-    //Setters
     public setEnvironment(environment: GlitterEnvironment): GlitterBridgeSDK {
-    //Set environment
         this._environment = environment;
 
         switch (environment) {
-        case GlitterEnvironment.mainnet:
-            this._bridgeConfig = BridgeMainnet;
-            break;
-        case GlitterEnvironment.testnet:
-            this._bridgeConfig = BridgeTestnet;
-            break;
-        default:
-            throw new Error("Environment not found");
+            case GlitterEnvironment.mainnet:
+                this._bridgeConfig = BridgeMainnet;
+                break;
+            case GlitterEnvironment.testnet:
+                this._bridgeConfig = BridgeTestnet;
+                break;
+            default:
+                throw new Error("Environment not found");
         }
-
-        //Get Tokens
-        const tokens: BridgeToken[] = [];
-        this._bridgeConfig.algorand.tokens.forEach((token) => {
-            tokens.push(token);
-        });
-        this._bridgeConfig.solana.tokens.forEach((token) => {
-            tokens.push(token);
-        });
-        BridgeTokens.loadConfig(tokens);
 
         return this;
     }
@@ -60,40 +46,40 @@ export class GlitterBridgeSDK {
         this._rpcOverrides[network] = rpc;
         return this;
     }
+    
     /**
-   * Initialize connections and SDK
-   * @param {BridgeNetworks[]} networks list of
-   * networks to connect to
-   * @returns {GlitterBridgeSDK}
-   */
+     * Initialize connections and SDK
+     * @param {BridgeNetworks[]} networks list of
+     * networks to connect to
+     * @returns {GlitterBridgeSDK}
+    */
     public connect(networks: BridgeNetworks[]): GlitterBridgeSDK {
         networks.forEach((network) => {
             /**
-       * TODO: Have a single method
-       * for each chain e.g
-       * interface ChainConnect { connect(network: BridgeNetworks): void; }
-       */
+             * TODO: Have a single method
+             * for each chain e.g
+             * interface ChainConnect { connect(network: BridgeNetworks): void; }
+            */
             switch (network) {
-            case BridgeNetworks.algorand:
-                this.connectToAlgorand();
-                break;
-            case BridgeNetworks.solana:
-                this.connectToSolana();
-                break;
-            case BridgeNetworks.Ethereum:
-                this.connectToEvmNetwork(BridgeNetworks.Ethereum);
-                break;
-            case BridgeNetworks.Polygon:
-                this.connectToEvmNetwork(BridgeNetworks.Polygon);
-                break;
-            case BridgeNetworks.Avalanche:
-                this.connectToEvmNetwork(BridgeNetworks.Avalanche);
-                break;
-            case BridgeNetworks.TRON:
-                this.connectToTron();
-                break;
-            }
-        });
+                case BridgeNetworks.algorand:
+                    this.connectToAlgorand();
+                    break;
+                case BridgeNetworks.solana:
+                    this.connectToSolana();
+                    break;
+                case BridgeNetworks.Ethereum:
+                    this.connectToEvmNetwork(BridgeNetworks.Ethereum);
+                    break;
+                case BridgeNetworks.Polygon:
+                    this.connectToEvmNetwork(BridgeNetworks.Polygon);
+                    break;
+                case BridgeNetworks.Avalanche:
+                    this.connectToEvmNetwork(BridgeNetworks.Avalanche);
+                    break;
+                case BridgeNetworks.TRON:
+                    this.connectToTron();
+                    break;
+            }});
 
         return this;
     }
@@ -104,20 +90,17 @@ export class GlitterBridgeSDK {
     private preInitializeChecks(network: BridgeNetworks) {
         if (!this._bridgeConfig) throw new Error("Glitter environment not set");
         /**
-     * TODO: have config keys in such
-     * a way that we directly check
-     * using JS this._bridgeConfig[network]
-     */
+         * TODO: have config keys in such
+         * a way that we directly check
+         * using JS this._bridgeConfig[network]
+         */
         const unavailableConfigError = `${network} Configuration unavailable`;
-        if (network === BridgeNetworks.TRON && !this._bridgeConfig.tron) {
-            throw new Error(unavailableConfigError);
-        } else if (
-            [
-                BridgeNetworks.Avalanche,
-                BridgeNetworks.Ethereum,
-                BridgeNetworks.Polygon,
-            ].includes(network) &&
-      !this._bridgeConfig.evm[network as BridgeEvmNetworks]
+        const isEvmInvalid = [BridgeNetworks.Avalanche, BridgeNetworks.Ethereum, BridgeNetworks.Polygon].includes(network) && !this._bridgeConfig.evm[network as BridgeEvmNetworks]
+        const isTronInvalid = network === BridgeNetworks.TRON && !this._bridgeConfig.tron
+        const isSolInvalid = network === BridgeNetworks.solana && !this._bridgeConfig.solana
+        const isAlgoInvalid = network === BridgeNetworks.algorand && !this._bridgeConfig.algorand
+        if (
+            isEvmInvalid || isTronInvalid || isSolInvalid || isAlgoInvalid
         ) {
             throw new Error(unavailableConfigError);
         }
@@ -131,48 +114,25 @@ export class GlitterBridgeSDK {
     }
 
     private connectToAlgorand(): GlitterBridgeSDK {
-    // Failsafe
-    // copy paste is an anti pattern
-    // https://en.wikipedia.org/wiki/Don%27t_repeat_yourself
-        if (!this._bridgeConfig) throw new Error("Glitter environment not set");
-        if (!this._bridgeConfig.algorand)
-            throw new Error("Algorand environment not set");
+        this.preInitializeChecks(BridgeNetworks.algorand)
 
         if (this._rpcOverrides[BridgeNetworks.algorand]) {
-            console.log(
-                "Algorand RPC override: " + this._rpcOverrides[BridgeNetworks.algorand]
-            );
-            this._bridgeConfig.algorand.serverUrl =
-        this._rpcOverrides[BridgeNetworks.algorand];
+            this._bridgeConfig!.algorand.serverUrl = this._rpcOverrides[BridgeNetworks.algorand];
         }
 
-        //Get the connections
-        this._algorand = new AlgorandConnect(this._bridgeConfig.algorand);
-
-        if (!this._algorand.client) throw new Error("Algorand client not set");
+        this._algorand = new AlgorandConnect(this._bridgeConfig!.algorand);
 
         return this;
     }
 
     private connectToSolana(): GlitterBridgeSDK {
-    //Failsafe
-    // same here
-        if (!this._bridgeConfig) throw new Error("Glitter environment not set");
-        if (!this._bridgeConfig.solana)
-            throw new Error("Solana environment not set");
+        this.preInitializeChecks(BridgeNetworks.solana);
 
         if (this._rpcOverrides[BridgeNetworks.solana]) {
-            console.log(
-                "Solana RPC override: " + this._rpcOverrides[BridgeNetworks.solana]
-            );
-            this._bridgeConfig.solana.server =
-        this._rpcOverrides[BridgeNetworks.solana];
+            this._bridgeConfig!.solana.server = this._rpcOverrides[BridgeNetworks.solana];
         }
 
-        this._solana = new SolanaConnect(this._bridgeConfig?.solana);
-        //(this._glitterNetwork.algorand.appProgramId);
-
-        if (!this._solana.client) throw new Error("Solana client not set");
+        this._solana = new SolanaConnect(this._bridgeConfig!.solana);
         return this;
     }
 
@@ -190,14 +150,12 @@ export class GlitterBridgeSDK {
         return this;
     }
 
-    /** Get chain specific connection */
-
     /**
-   * Returns EVMConnect for
-   * a specific evm network
-   * @param {BridgeEvmNetworks} network
-   * @returns {EvmConnect | undefined}
-   */
+     * Returns EVMConnect for
+     * a specific evm network
+     * @param {BridgeEvmNetworks} network
+     * @returns {EvmConnect | undefined}
+     */
     public getEvmNetwork(evmNetwork: BridgeEvmNetworks): EvmConnect | undefined {
         return this._evm.get(evmNetwork);
     }
